requirements.txt
Flask>=2.3.0
flask-sqlalchemy>=3.0.0
python-dotenv>=1.0.0
groq>=0.4.0
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
accelerate>=0.25.0
safetensors>=0.4.0
